attach command works. added a lot of prints to track events. start mostly works. I think we need to send each HELLO in its own thread otherwise it seems to block after the first HELLO. Other than that everything seems to work.

+++ src/main/java/socs/network/message/SOSPFPacket.java
@@ -8,14 +8,14 @@ public class SOSPFPacket implements Serializable {
 
   //for inter-process communication
   public String srcProcessIP;
-  public short srcProcessPort;
+  public int srcProcessPort;
 
   //simulated IP address
   public String srcIP;
   public String dstIP;
 
   //common header
-  public short sospfType; //0 - HELLO, 1 - LinkState Update
+  public int sospfType; //0 - HELLO, 1 - LinkState Update
   public String routerID;
 
   //used by HELLO message to identify the sender of the message
